requirements.txt
Django==5.2.2
django-jazzmin==3.0.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
django-redis==5.4.0
dnspython==2.6.1
pymongo==4.7.2
gunicorn
fastapi
uvicorn
aiomysql
asyncio